Make imputation reason code static table consistent with other static tables
